requirements.txt
streamlit
langchain
openai
qdrant-client
sentence-transformers
beautifulsoup4
trafilatura
python-dotenv
google-generativeai
langchain_qdrant